codegen: relocate Soroban target into targets/soroban/
Move soroban.rs, dispatch/soroban.rs, encoding/soroban_encoding.rs and
events/soroban.rs under targets/soroban/ as history-preserving renames,
and move the SorobanTarget impl alongside them. Repoint imports and bump
the cross-boundary helpers to pub(crate). Pure relocation: golden CFG
tests pass with zero CHECK edits.
