+++ src/codegen/targets/soroban/encoding.rs
